Add CreateNewLibrary Transaction Definition
This commit introduces the CreateNewLibrary transaction definition. The transaction includes details such as the tag createNewLibrary, label Create New Library, description Create a New Library, method POST, callers limited to admin users from org3MSP and orgMSP, and an argument name which is required. The routine implements the logic for creating and saving the library asset in the ledger, and emitting an event upon successful creation. This enhances the chaincode's functionality by providing a structured way to handle the creation of libraries, ensuring that the transaction is properly defined and executed.